requirements.txt
tk==0.1.0
tabula-py==2.7.0
pillow==10.0.1
